ivi-controller: introduced ivi-controller header file
New header file for ivi-controller with ivishell and extended ivisurface to replace surface_ctx from ivi-input-controller.
